+++ calcipy/tasks/types.py
@@ -1,3 +1,4 @@
+# noqa: A005
 """Types CLI."""
 
 from invoke.context import Context