+++ tests/tasks/test_lint.py
@@ -11,7 +11,7 @@ from calcipy.tasks.lint import ALL_PRE_COMMIT_HOOK_STAGES, check, fix, pre_commi
     [
         (check, {}, [f'{python_m()} ruff check "./calcipy" ./tests']),
         (fix, {}, [f'{python_m()} ruff check "./calcipy" ./tests --fix']),
-        (watch, {}, [f'{python_m()} ruff check "./calcipy" ./tests --watch --show-source']),
+        (watch, {}, [f'{python_m()} ruff check "./calcipy" ./tests --watch']),
         (pre_commit, {}, [
             call('which pre-commit', warn=True, hide=True),
             'pre-commit install',